+++ app/Http/Controllers/MenuController.php
@@ -3,9 +3,11 @@
 namespace App\Http\Controllers;
 
 use App\Models\Menu;
-use App\Models\JenisMenu;
 use App\Models\Tenant;
+use App\Models\JenisMenu;
+use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Auth;
 
 
@@ -13,30 +15,76 @@ class MenuController extends Controller
 {
     public function index()
     {
-        $menu = Menu::getAllMenu();
-        return view("admin.menu.index", ['data' => $menu]);
+        if (Auth::check()) {
+            $id = Auth::id();
+            $usr = User::find($id);
+            $type = $usr->id_type_user;
+            if ($type === "tnt") {
+                $tenant = Tenant::where('id_user', $id)->first();
+                $idTenant = $tenant->id_tenant;
+            }
+            $menu = DB::table('tbl_menu as a')
+                ->join('mst_jenis_menu as b', 'a.id_jenis_menu', '=', 'b.id_jenis_menu')
+                ->join('mst_tenant as c', 'a.id_tenant', '=', 'c.id_tenant')
+                ->select('a.*', 'b.*', 'c.*');
+            if ($type === "tnt") {
+                $menu->where('c.id_tenant', $idTenant);
+            }
+            $result = $menu->paginate(10);;
+            // $menu = Menu::getAllMenu();
+            return view("admin.menu.index", ['data' => $result]);
+        }
+        return view('login.index');
     }
-
 
 
     public function tambah()
     {
-        $tenant = Tenant::get();
-        $jenisMenu = JenisMenu::get();
-        return view("admin.menu.form",['jenisMenu'=>$jenisMenu, 'tenant' => $tenant]);
+        if (Auth::check()) {
+            $id = Auth::id();
+            $usr = User::find($id);
+            $type = $usr->id_type_user;
+            if ($type === "tnt") {
+                $tenant = Tenant::where('id_user', $id)->first();
+                $idTenant = $tenant->id_tenant;
+                $tenant = Tenant::where('id_tenant',$idTenant)->get();
+            }else{
+                $tenant = Tenant::get();
+            }
+            $jenisMenu = JenisMenu::get();
+            return view("admin.menu.form", ['jenisMenu' => $jenisMenu, 'tenant' => $tenant]);
+        }
+        return view('login.index');
     }
 
-    public function edit($id)
+    public function edit($idMenu)
     {
-        $tenant = Tenant::get();
-        $jenisMenu = JenisMenu::get();
-        $idMenu = Menu::find($id);
-        return view('admin.menu.form', ['menu' => $idMenu,'jenisMenu' => $jenisMenu, 'tenant' => $tenant]);
+        if (Auth::check()) {
+            $id = Auth::id();
+            $usr = User::find($id);
+            $type = $usr->id_type_user;
+            if ($type === "tnt") {
+                $tenant = Tenant::where('id_user', $id)->first();
+                $idTenant = $tenant->id_tenant;
+                $tenant = Tenant::where('id_tenant', $idTenant)->get();
+            } else {
+                $tenant = Tenant::get();
+            }
+            $jenisMenu = JenisMenu::get();
+            $menu = Menu::find($idMenu);
+            return view('admin.menu.form', ['menu' => $menu, 'jenisMenu' => $jenisMenu, 'tenant' => $tenant]);
+        }
+        return view('login.index');
     }
 
     public function update($id, Request $request){
+        if (Auth::check()) {
         $request->validate([
-            'nama_gambar' => 'nullable|image|mimes:jpeg,png,jpg,gif|max:2048', // Tetapkan nullable untuk tidak wajib mengunggah file baru
+            'nama_gambar' => 'nullable|image|mimes:jpeg,png,jpg,gif|max:2048',
+            'nama_menu' => ['required', 'max:35'],
+            'id_jenis_menu' => 'required',
+            'harga_jual' => ['required', 'max:6'],
+            'id_tenant' => 'required',
         ]);
         $data = [
             'nama_menu' => $request->nama_menu,
@@ -50,9 +98,9 @@ class MenuController extends Controller
             if ($image->getError() == UPLOAD_ERR_OK) {
                 $imageName = time() . '.' . $image->getClientOriginalExtension();
                 $image->move(public_path('images'), $imageName);
-                $imagePath = 'images/' . $imageName;
-                $imageUrl = asset($imagePath);
-                $data['nama_gambar'] = $imageUrl;
+               // $imagePath = 'images/' . $imageName;
+           //     $imageUrl = asset($imagePath);
+                $data['nama_gambar'] = $imageName;
             } else {
                 // Handle upload error
                 return redirect()->back()->with('error', 'Failed to upload gambar.');
@@ -61,13 +109,19 @@ class MenuController extends Controller
 
         Menu::find($id)->update($data);
         return redirect()->route('menu');
-
+        }
+        return view('login.index');
     }
 
 
     public function simpan(Request $request){
+        if (Auth::check()) {
         $request->validate([
-            'nama_gambar' => 'required|image|mimes:jpeg,png,jpg,gif|max:2048', // max 2MB
+            'nama_gambar' => 'required|image|mimes:jpeg,png,jpg,gif|max:2048',
+            'nama_menu'=>['required','max:35'],
+            'id_jenis_menu'=> 'required',
+            'harga_jual'=>['required', 'max:6'],
+            'id_tenant'=> 'required',
         ]);
 
         if ($request->hasFile('nama_gambar')) {
@@ -77,8 +131,8 @@ class MenuController extends Controller
                 // Process and save the file
                 $imageName = time() . '.' . $image->getClientOriginalExtension();
                 $image->move(public_path('images'), $imageName);
-                $imagePath = 'images/' . $imageName;
-                $imageUrl = asset($imagePath);
+               // $imagePath = 'images/' . $imageName;
+               // $imageUrl = asset($imagePath);
 
                 $data = [
                     'nama_menu' => $request->nama_menu,
@@ -86,7 +140,7 @@ class MenuController extends Controller
                     'harga_jual' => $request->harga_jual,
                     'id_tenant' => $request->id_tenant,
                     'status_menu' => $request->status_menu,
-                    'nama_gambar' => $imageUrl
+                    'nama_gambar' => $imageName
                 ];
 
                 Menu::create($data);
@@ -98,13 +152,17 @@ class MenuController extends Controller
         } else {
             return redirect()->back()->with('error', 'No gambar uploaded.');
         }
-
+}
+        return view('login.index');
     }
 
     public function hapus($id_menu)
     {
+        if (Auth::check()) {
         Menu::find($id_menu)->delete();
         return redirect()->route('menu');
+        }
+        return view('login.index');
     }
 
 
@@ -121,6 +179,7 @@ class MenuController extends Controller
                 'menu' => $menu
             ]);
         }
+        return view('login.index');
     }
